requirements.txt
Flask==1.0.2
hunspell==0.5.5
numpy==1.16.2
editdistance==0.5.2
gensim==3.7.1
Keras==2.2.4
nltk==3.4
pandas==0.24.1
scikit-learn==0.21.3
spacy==2.1.8
tensorflow-gpu==1.14.0
